pyproject.toml
[tool.poetry]
name = "ninja-backend"
version = "0.1.0"
description = "Backend server of NINJA project"
authors = ["Ben Chen <chanben04gz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
sqlalchemy = "^2.0.35"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
python-dotenv = "^1.0.1"
flask-cors = "^5.0.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
